# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Movie API",
        "description": "API for managing films, actors, critics, and users",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/films/{id}/actors": {
            "get": {
                "tags": [
                    "Actors"
                ],
                "summary": "Get all actors for a specific film",
                "operationId": "188c4bc600d10f1cd15c7e131ff7607e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of actors in the film"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/api/critics/{id}": {
            "delete": {
                "tags": [
                    "Critics"
                ],
                "summary": "Delete a critic review",
                "operationId": "4c7bbd907e57760bbbfc560c1ba8e1ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Critic ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Critic deleted"
                    },
                    "404": {
                        "description": "Critic not found"
                    }
                }
            }
        },
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get all films",
                "operationId": "271e474415d2f29bf3c163247c5f4639",
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/films/{id}/critics": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get a film along with its critics",
                "operationId": "13cb33623e1a120faf393524ead1f850",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film details with critics"
                    }
                }
            }
        },
        "/api/films/{id}/average-score": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get the average score for a film",
                "operationId": "8aef37460ef035739f58ccd5343f0618",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Average film score"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "email",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Updated email",
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "description": "Updated first name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Updated last name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    }
                }
            }
        },
        "/api/users/{id}/preferred-language": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a user's preferred language",
                "operationId": "b0616f0bd4be35487956ca77ee06a680",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's preferred language"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Film": {
                "title": "Film",
                "description": "Film model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "ACADEMY DINOSAUR"
                    },
                    "description": {
                        "type": "string",
                        "example": "A Epic Drama of a Feminist And a Mad Scientist who must Battle a Teacher in The Canadian Rockies"
                    },
                    "release_year": {
                        "type": "integer",
                        "example": 2006
                    },
                    "language_id": {
                        "description": "Foreign key referring to Language",
                        "type": "integer",
                        "example": 1
                    },
                    "length": {
                        "description": "Film length in minutes",
                        "type": "integer",
                        "example": 86
                    },
                    "rating": {
                        "type": "string",
                        "example": "PG"
                    },
                    "special_features": {
                        "type": "string",
                        "example": "Deleted Scenes,Behind the Scenes"
                    },
                    "image": {
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2006-02-15 15:03:42"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Actors",
            "description": "Actors"
        },
        {
            "name": "Critics",
            "description": "Critics"
        },
        {
            "name": "Films",
            "description": "Films"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}